Strip out deprecated workload identity endpoints & configuration
